Redesign mocking
The previous implementation of mocking was rather tricky because
an error to mock in a test causes issues to subsequent tests.

Also, parallelism (disabled by default) does not cope well with
constantly changing global variables.

On the contrary, setting the dependencies into structs, we really
have no issues and in fact the coverage now is better [*]

[*] There was something wrong in the previous implementation and
the coverage was acually lower than I thought.